Add support for Scala 2.13 (#290)

+++ src/test/scala/treadle/DynamicMemorySearch.scala
@@ -71,7 +71,7 @@ class DynamicMemorySearch extends AnyFreeSpec with Matchers with LazyLogging {
         tester.poke("io_en", 0)
       }
 
-      def checkSearch(searchValue: BigInt): Unit = {
+      def checkSearch(searchValue: Int): Unit = {
         val expectedIndex = list.indexOf(searchValue) match {
           case -1 => list.length
           case x  => x
